ai-engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.0
sentence-transformers==2.2.2
spacy==3.7.2
torch==2.1.0
transformers==4.35.0
